requirements.txt
aiohttp==3.8.3
arrow==1.2.3
cachetools==5.2.0
cryptography==38.0.3
cssselect==1.2.0
discord.py[voice]==2.1.0
ftfy==6.1.1
geopy==2.3.0
humanfriendly==10.0
lxml==4.9.1
markovify==0.9.4
motor==3.1.1
pillow==9.3.0
psutil==5.9.4
pymongo==4.3.3
python-aiml==0.9.3
pyyaml==6.0
redis==4.3.4
translate==3.6.1
twitchio==2.5.0
youtube_dl>=2021.12.17
